Add files via upload
I conducted the data analysis that will help to learn, how Netflix is successful in content release and types of content, mostly released, successful genre content, Successful pair of Actors, and director movie released mostly. When analyzing the data, we can identify, and get to know the data in past, and present that helps to predict future trends.